+++ spec/integration/spec_file_load_errors_spec.rb
@@ -31,6 +31,29 @@ RSpec.describe 'Spec file load errors' do
     end
   end
 
+  it 'nicely handles load-time errors from --require files' do
+    write_file_formatted "helper_with_error.rb", "raise 'boom'"
+
+    run_command "--require ./helper_with_error"
+    expect(last_cmd_exit_status).to eq(failure_exit_code)
+    output = normalize_durations(last_cmd_stdout)
+    expect(output).to eq unindent(<<-EOS)
+
+      An error occurred while loading ./helper_with_error.
+      Failure/Error: raise 'boom'
+
+      RuntimeError:
+        boom
+      # ./helper_with_error.rb:1#{spec_line_suffix}
+      No examples found.
+
+
+      Finished in n.nnnn seconds (files took n.nnnn seconds to load)
+      0 examples, 0 failures, 1 error occurred outside of examples
+
+    EOS
+  end
+
   it 'nicely handles load-time errors in user spec files' do
     write_file_formatted "1_spec.rb", "
       boom